Add utility removeRequireObjectProperty
